requirements.txt
beautifulsoup4
Django
djangorestframework
factory-boy
mock
numpy
pytz
